testing fix. I missed SpringBootTest annotation on SimplJavaTest so, to get it work I renamed root package for tests plus additional props in app.props for REST
