+++ src/main/resources/application.properties
@@ -2,3 +2,4 @@ calculation.avg = true
 calculation.sum = true
 calculation.min = true
 calculation.max = true
+console = true
