+++ src/test/java/calculations/SimplJavaTest.java
@@ -1,9 +1,11 @@
-package testing;
+package calculations;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
 @ActiveProfiles("test")
+@SpringBootTest
 public class SimplJavaTest {
 
     @Test
