+++ src/test/resources/application.properties
@@ -2,3 +2,4 @@ calculation.avg = true
 calculation.sum = true
 calculation.min = false
 calculation.max = true
+console = false
